Fixed duplicate prefix caused by reloading gossipframe
This happens a lot when you interact with an NPC that has several options offered including one or more quest. If you accept a new quest on the NPC and it returns you to the start of the GossipFrame, it would append another number prefix that isn't guaranteed to be the same as the original one.

An example of this is during the blue dragonflight quest line in Thaldrazsus. After accepting quest 73069 (Sindragosa and Malygos's Rest), Senegos would have duplicated prefixes.